Changing to us_housing_units_completed
It's not a big deal, but the current version of the tutorial is a mix of using the tables <code>us_housing_units</code> and <code>us_housing_units_completed</code>. Perhaps this was part of the tutorial. But in case not, I have made an attempt here to change all text and links to use the <code>_completed</code> table.